+++ platform/flowglad-next/src/app/store/products/[id]/prices/data-table.tsx
@@ -1,0 +1,233 @@
+'use client'
+
+import * as React from 'react'
+import {
+  ColumnFiltersState,
+  SortingState,
+  VisibilityState,
+  flexRender,
+  getCoreRowModel,
+  getFilteredRowModel,
+  getPaginationRowModel,
+  getSortedRowModel,
+  useReactTable,
+} from '@tanstack/react-table'
+import { Input } from '@/components/ui/input'
+import { Button } from '@/components/ui/button'
+import {
+  Table,
+  TableBody,
+  TableCell,
+  TableHead,
+  TableHeader,
+  TableRow,
+} from '@/components/ui/table'
+import { DataTableViewOptions } from '@/components/ui/data-table-view-options'
+import { DataTablePagination } from '@/components/ui/data-table-pagination'
+import { columns, PriceTableRowData } from './columns'
+import { usePaginatedTableState } from '@/app/hooks/usePaginatedTableState'
+import { trpc } from '@/app/_trpc/client'
+import debounce from 'debounce'
+import { Search, Plus } from 'lucide-react'
+import { PriceType } from '@/types'
+
+export interface PricesTableFilters {
+  productId?: string
+  type?: PriceType
+  isDefault?: boolean
+}
+
+interface PricesDataTableProps {
+  productId: string
+  filters?: PricesTableFilters
+  onCreatePrice?: () => void
+}
+
+export function PricesDataTable({
+  productId,
+  filters = {},
+  onCreatePrice,
+}: PricesDataTableProps) {
+  // Server-side filtering (preserve enterprise architecture)
+  const [inputValue, setInputValue] = React.useState('')
+  const [searchQuery, setSearchQuery] = React.useState('')
+  const debouncedSetSearchQuery = debounce(setSearchQuery, 1000)
+
+  // Page size state for server-side pagination
+  const [currentPageSize, setCurrentPageSize] = React.useState(10)
+
+  React.useEffect(() => {
+    debouncedSetSearchQuery(inputValue)
+  }, [inputValue, debouncedSetSearchQuery])
+
+  const {
+    pageIndex,
+    pageSize,
+    handlePaginationChange,
+    data,
+    isLoading,
+    isFetching,
+  } = usePaginatedTableState<PriceTableRowData, PricesTableFilters>({
+    initialCurrentCursor: undefined,
+    pageSize: currentPageSize,
+    filters: {
+      ...filters,
+      productId,
+    },
+    searchQuery: searchQuery,
+    useQuery: trpc.prices.getTableRows.useQuery,
+  })
+
+  // Client-side features (Shadcn patterns)
+  const [sorting, setSorting] = React.useState<SortingState>([])
+  const [columnFilters, setColumnFilters] =
+    React.useState<ColumnFiltersState>([])
+  const [columnVisibility, setColumnVisibility] =
+    React.useState<VisibilityState>({})
+
+  const table = useReactTable({
+    data: data?.items || [],
+    columns,
+    manualPagination: true, // Server-side pagination
+    manualSorting: false, // Client-side sorting on current page
+    manualFiltering: false, // Client-side filtering on current page
+    pageCount: Math.ceil((data?.total || 0) / currentPageSize),
+    onSortingChange: setSorting,
+    onColumnFiltersChange: setColumnFilters,
+    onColumnVisibilityChange: setColumnVisibility,
+
+    // CRITICAL: Bridge TanStack Table pagination to server-side pagination
+    onPaginationChange: (updater) => {
+      const newPagination =
+        typeof updater === 'function'
+          ? updater({ pageIndex, pageSize: currentPageSize })
+          : updater
+
+      // Handle page size changes
+      if (newPagination.pageSize !== currentPageSize) {
+        setCurrentPageSize(newPagination.pageSize)
+        handlePaginationChange(0) // Reset to first page
+      }
+      // Handle page navigation
+      else if (newPagination.pageIndex !== pageIndex) {
+        handlePaginationChange(newPagination.pageIndex)
+      }
+    },
+
+    getCoreRowModel: getCoreRowModel(),
+    getSortedRowModel: getSortedRowModel(),
+    getFilteredRowModel: getFilteredRowModel(),
+
+    // CRITICAL: Use dynamic page size in state
+    state: {
+      sorting,
+      columnFilters,
+      columnVisibility,
+      pagination: { pageIndex, pageSize: currentPageSize },
+    },
+  })
+
+  return (
+    <div className="w-full">
+      {/* Enhanced toolbar with all improvements */}
+      <div className="flex items-center py-4">
+        <div className="relative">
+          <Search className="absolute left-3 top-1/2 -translate-y-1/2 h-4 w-4 text-muted-foreground" />
+          <Input
+            placeholder="Search prices..."
+            value={inputValue}
+            onChange={(e) => setInputValue(e.target.value)}
+            className="max-w-sm pl-9"
+            disabled={isLoading}
+          />
+          {isFetching && (
+            <div className="absolute right-3 top-1/2 -translate-y-1/2">
+              <div className="h-4 w-4 animate-spin rounded-full border-2 border-gray-300 border-t-gray-600"></div>
+            </div>
+          )}
+        </div>
+        <div className="flex items-center gap-2 ml-auto">
+          <DataTableViewOptions table={table} />
+          {onCreatePrice && (
+            <Button onClick={onCreatePrice}>
+              <Plus className="w-4 h-4 mr-2" />
+              Create Price
+            </Button>
+          )}
+        </div>
+      </div>
+
+      {/* Table */}
+      <div className="rounded-md border">
+        <Table>
+          <TableHeader>
+            {table.getHeaderGroups().map((headerGroup) => (
+              <TableRow key={headerGroup.id}>
+                {headerGroup.headers.map((header) => {
+                  return (
+                    <TableHead
+                      key={header.id}
+                      style={{ width: header.getSize() }}
+                    >
+                      {header.isPlaceholder
+                        ? null
+                        : flexRender(
+                            header.column.columnDef.header,
+                            header.getContext()
+                          )}
+                    </TableHead>
+                  )
+                })}
+              </TableRow>
+            ))}
+          </TableHeader>
+          <TableBody>
+            {isLoading ? (
+              <TableRow>
+                <TableCell
+                  colSpan={columns.length}
+                  className="h-24 text-center text-muted-foreground"
+                >
+                  Loading...
+                </TableCell>
+              </TableRow>
+            ) : table.getRowModel().rows?.length ? (
+              table.getRowModel().rows.map((row) => (
+                <TableRow
+                  key={row.id}
+                  className={isFetching ? 'opacity-50' : ''}
+                >
+                  {row.getVisibleCells().map((cell) => (
+                    <TableCell
+                      key={cell.id}
+                      style={{ width: cell.column.getSize() }}
+                    >
+                      {flexRender(
+                        cell.column.columnDef.cell,
+                        cell.getContext()
+                      )}
+                    </TableCell>
+                  ))}
+                </TableRow>
+              ))
+            ) : (
+              <TableRow>
+                <TableCell
+                  colSpan={columns.length}
+                  className="h-24 text-center text-muted-foreground"
+                >
+                  No results.
+                </TableCell>
+              </TableRow>
+            )}
+          </TableBody>
+        </Table>
+      </div>
+
+      {/* Enhanced pagination with proper spacing */}
+      <div className="py-2">
+        <DataTablePagination table={table} totalCount={data?.total} />
+      </div>
+    </div>
+  )
+}
